story update
