--- packages/react-components/react-timepicker-compat-preview/stories/TimePicker/TimePickerCustomParsingAndValidation.md
@@ -1,4 +1,0 @@
-V9 TimePicker should be used together with `Field` component, and it provides more flexibility for custom validation.
-
-- You can perform custom parsing and validation for freeform input using `formatDateToTimeString`.
-- Validation of the selected time option from the dropdown can be achieved by validating the `selectedTime` within `onTimeChange` callback.
