+++ packages/react-components/react-timepicker-compat-preview/stories/TimePicker/TimePickerFreeformCustomParsing.md
@@ -1,0 +1,4 @@
+This story sets custom time string in the dropdown options, and performs custom parsing from the input text to Date object on selection.
+
+- The time display format in the dropdown can be customized using `formatDateToTimeString`.
+- Freefrom TimePicker can have custom parsing for input text to Date object using `parseTimeStringToDate`.
