+++ packages/react-components/react-timepicker-compat-preview/stories/TimePicker/TimePickerFreeformWithErrorHandling.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Field, FieldProps, makeStyles } from '@fluentui/react-components';
 import { TimePicker, TimePickerErrorType, TimePickerProps } from '@fluentui/react-timepicker-compat-preview';
-import story from './TimePickerFreeform.md';
+import story from './TimePickerFreeformWithErrorHandling.md';
 
 const useStyles = makeStyles({
   control: {
